Split 'TypesCollection' in separate AO/AE files
